+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] - 2020-04-10
+
+### Added
+
+-   support for `beforeEach` and `afterEach` hooks
+-   improved log message for `cy.request`
+
 ## [1.1.2] - 2020-04-07
 
 ### Fixed